Aggiornato link eliminazione solo per amministratore

+++ WebContent/object/routePage.jsp
@@ -55,7 +55,9 @@
 	
 	<br>
 	
-	<a href="<%=request.getContextPath().toString()%>/deleteRoute.jsp?idr=<%=route.getId()%>">Elimina itinerario</a>
+	<%if(u.getAuthorization() > 0){ %>
+		<a href="<%=request.getContextPath().toString()%>/deleteRoute.jsp?idr=<%=route.getId()%>">Elimina itinerario</a>
+	<%} %>
 	<script type="text/javascript">
 		function searchLocation(){
 			var xhttp;
